fix(neuron-ui): update interface ContentProps cause actionCreators not passed to children, seperate reducer of Header
